Changed to get S3 IDs directly from bucket

+++ dcs/core/data_repository.py
@@ -19,6 +19,7 @@ import re
 from datetime import timedelta
 
 from ghga_service_commons.utils import utc_dates
+from hexkit.protocols.dao import MultipleHitsFoundError, NoHitsFoundError
 from pydantic import BaseSettings, Field, PositiveInt, validator
 
 from dcs.adapters.outbound.http import exceptions
@@ -158,7 +159,6 @@ class DataRepository(DataRepositoryPort):
 
         # Successfully staged, update access information now
         drs_object_with_access_time.last_accessed = utc_dates.now_as_utc()
-        drs_object_with_access_time.in_outbox = True
         try:
             await self._drs_object_dao.update(drs_object_with_access_time)
         except ResourceNotFoundError as error:
@@ -184,24 +184,33 @@ class DataRepository(DataRepositoryPort):
         is met or exceeded, the corresponding file is removed from the outbox.
         """
         threshold = utc_dates.now_as_utc() - timedelta(days=cache_timeout)
+
         # filter to get all files in outbox that should be removed
-        mapping = {"and": [{"last_accessed": {"lte": threshold}}, {"in_outbox": True}]}
-        async for expired_file in self._drs_object_dao.find_all(mapping=mapping):
-            s3_id = expired_file.file_id
-
+        outbox_ids = await self._object_storage.list_all_object_ids(
+            bucket_id=self._config.outbox_bucket
+        )
+        for outbox_id in outbox_ids:
+            mapping = {"file_id": outbox_id}
             try:
-                await self._object_storage.delete_object(
-                    bucket_id=self._config.outbox_bucket, object_id=s3_id
-                )
-            except (
-                self._object_storage.ObjectError,
-                self._object_storage.ObjectStorageProtocolError,
-            ) as error:
-                raise self.CleanupError(object_id=s3_id, from_error=error) from error
+                drs_object = await self._drs_object_dao.find_one(mapping=mapping)
+            except (MultipleHitsFoundError, NoHitsFoundError) as error:
+                raise self.CleanupError(
+                    object_id=outbox_id, from_error=error
+                ) from error
 
-            # update database entry
-            expired_file.in_outbox = False
-            self._drs_object_dao.upsert(expired_file)
+            # only remove file if last access is later than cache timeout days ago
+            if drs_object.last_accessed <= threshold:
+                try:
+                    await self._object_storage.delete_object(
+                        bucket_id=self._config.outbox_bucket, object_id=outbox_id
+                    )
+                except (
+                    self._object_storage.ObjectError,
+                    self._object_storage.ObjectStorageProtocolError,
+                ) as error:
+                    raise self.CleanupError(
+                        object_id=outbox_id, from_error=error
+                    ) from error
 
     async def register_new_file(self, *, file: models.DrsObject):
         """Register a file as a new DRS Object."""
@@ -209,7 +218,6 @@ class DataRepository(DataRepositoryPort):
         file_with_access_time = models.AccessTimeDrsObject(
             **file.dict(),
             last_accessed=utc_dates.now_as_utc(),
-            in_outbox=False,
         )
         # write file entry to database
         await self._drs_object_dao.insert(file_with_access_time)
